api/requirements.txt
# Road-score API (run: uvicorn api.main:app --reload)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Local Whisper transcription for /v1/media/audio-score (transcribe-then-score path)
faster-whisper>=1.0.0
